Move PluginMessage to tensai-api
